Adds support for configuring NIC to OVS
OVN-k8s needs to be able to move the primary NIC of the host into the
OVS bridge, making the physical NIC essentially a Layer 2 port. This
patches uses nmcli to configure OVS with the NIC on it, as well as bring
up the Layer 3 OVS interface. The behavior added by this patch selects
the interface based on the default route found on the host.

Follow up patches will address bonding and other types of configuration.

Signed-off-by: Tim Rozet <trozet@redhat.com>